+++ crates/maple_cli/src/stdio_server/mod.rs
@@ -94,6 +94,7 @@ fn loop_handle_rpc_message(rx: &Receiver<String>) {
             match &msg.method[..] {
                 "initialize_global_env" => initialize_global(msg), // should be called only once.
                 "init_ext_map" => message_handlers::parse_filetypedetect(msg),
+                "preview/file" => message_handlers::preview_file(msg),
                 "filer" => filer::handle_filer_message(msg),
                 "quickfix" => quickfix::preview_quickfix_entry(msg),
                 "dumb_jump" => dumb_jump::handle_dumb_jump_message(msg),