+++ crates/maple_cli/src/stdio_server/session/message_handlers/mod.rs
@@ -2,9 +2,10 @@
 
 use std::collections::HashMap;
 
+use anyhow::Result;
 use serde_json::json;
 
-use crate::stdio_server::{types::Message, write_response};
+use crate::stdio_server::{previewer, types::Message, write_response};
 
 pub fn parse_filetypedetect(msg: Message) {
     let output = msg.get_string_unsafe("autocmd_filetypedetect");
@@ -29,4 +30,26 @@ pub fn parse_filetypedetect(msg: Message) {
         json!({ "id": msg.id, "force_execute": true, "result": json!({"ext_map": ext_map}) });
 
     write_response(result);
+}
+
+async fn preview_file_impl(msg: Message) -> Result<()> {
+    let fpath = msg.get_string("fpath")?;
+    let winwidth = msg.get_u64("preview_width")?;
+    let winheight = msg.get_u64("preview_height")?;
+
+    let (lines, fname) = previewer::preview_file(fpath, winheight as usize, winwidth as usize)?;
+
+    let result = json!({"id": msg.id, "result": json!({"lines": lines, "fname": fname})});
+
+    write_response(result);
+
+    Ok(())
+}
+
+pub fn preview_file(msg: Message) {
+    tokio::spawn(async move {
+        if let Err(e) = preview_file_impl(msg).await {
+            log::error!("Error when previewing the file: {}", e);
+        }
+    });
 }