+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [neovim] Add zindex option to fix the tricky floating_win overlapping, and add border for the preview window, use `let g:clap_popup_border = 'nil'` to disable the order. #693
 - Impl preview for `quickfix` provider. #691
+- Impl `preview/file` for easier external async preview integration. #706
 
 ### Changed
 
@@ -17,6 +18,7 @@
 - [neovim] Fix the action dialog creation using floating_win. #688
 - Fix the indicator winwidth is not flexible. #687
 - Fix the icon offset when restoring the full display line for grep provider. #701
+- Fix the Pyo3 compilation on M1. #707
 
 ### Perf